requirements.txt
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
urllib3==1.21.1
websocket-client==0.44.0
pyee==3.0.
